Final commit of 2D read before merging into master branch.
OVERVIEW:
    The HDF files on the BEHR website, as well as the .mat files
Ashley left in the "NEWER" folder (on the old file server) have the
variables stored in 2D arrays, which retains the spatial orientation
of the OMI pixels. The arrays are n x 60, where n is the number of
pixels along track, and 60 is the number of rows across track.
    In the copies of the code that I'd originally found, any pixel
outside the specified lat/lon boundaries was discarded.  This
necessitated the reduction of the variables to 1D, since there was no
guarentee that all pixels in a given row would be kept. To keep the 2D
structure, I've set it up now so that the pixels are rejected on a
row-by-row basis, and if any row has even a single pixel inside in
lat/lon bounds, the entire row is kept.

MODIFIED FILES:
    omiAmfAK2 - removed the output of the AMF calculated from the
average scattering weights. Commented out the calculation of that AMF.

    BEHR_main - generally changed it to Data structures with 2D
variables. Specifically:
        1) Variables read in from Data structure are converted to
column vectors (lon, lat, sza, vza, phi, surfPres, albedo, cldPres,
cldFrac, cldRadFrac)
        2) Variables (like mon, cloudalbedo) that are created to be
the same size as the Data variables now transform those variables into
column vectors and then find their size.
        3) The AMF, scattering weights, and averaging kernels are now
reshaped to 2- or 3-D (whichever is appropriate for that variable)
before saving to the BEHR Data structure.
        Also removed all references to the testing variable, AMF2.

    BEHR_hdf_v2 - added units for BEHRColumnAmountNO2Trop_L3 and
BEHRColumnAmount_L3MODISCloud in the attributes table.  Also now
checks if the swath_id value (maximum of all values of Swath) is a NaN
- a necessary change now that fill values are left in the OMI
  structure instead of zeroes.

    read_omno2_v_aug2012 - here is where the biggest changes had to be
made to keep the variables 2D.
        1) Time now needs to be kept in its original orientation
before being replicated over all rows.
        2) The initial rejection of pixels based on the lat/lon
boundaries (around line 365) rejects row-by-row, rather than removing
pixels individually.
        3) Lon & latcorn now need to be rearranged so that the corners
are the first dimension, and the second and third dimensions
correspond to the the 2D dimensions of the other variables.
        4) After all variables are loaded, across track lines
containing no useful pixels are removed. The exact same logical
indices from ~line 365 are used here directly.
        5) When the variables are saved to the Data structure, the (:)
indexing is removed - this was used before to ensure that the
variables went in as column vectors.
        6) MODIS Cloud data now loops over number of elements in
Data.Latitude rather than length - length will not capture a 2D array.
Linear indexing now used in assigning cloud values. MODIS Albedo and
GLOBE terrain pressure did not need changed.

        **There were some other, unrelated changes too:
        I) Deleted unnecessary code, including the old way of
formatting the day of year string and the GLOBEHeight variables.
        II) MODIS cloud now uses the function behr_fill_val to
generate the fill value included if MODIS cloud data is not available.

    hdf_quadrangle_BEHR - removed the BEHRColumnAmountNO2Trop_AMF2
field (debugging, not needed). Also switched the isnan checks before
gridding to use BEHRColumnAmountNO2Trop_val rather than the variable
directly from Data. The _val version already handles the 2D nature of
the variable correctly, no need to reproduce that elsewhere.