Automatic satellite data downloading seems to be working.
NEW FILES:

    Downloading/check_myd06_files.m - function that tries to load every MYD06
file after a certain date (currently hard coded). If any error for hdfinfo(), it
prints that to a file in the home directory. This was needed to find a few
corrupt MYD06 files which were causing BEHR to abort during
read_omno2_v_aug2012.

    Downloading/get_omno2.sh - bash script that will wait for a delivery notice
to be placed in the OMNO2/download_staging folder, then find all files in that
notice not already present, then download and sort them.

    Downloading/order_omno2.sh - bash script that requests the delivery notice
from NASA, waits for it to be delivered to the nasa user, then copies it to the
required locations.  Should be run as root, since it has to be able to get into
the nasa home folder.

    Downloading/run_read_omno2.sh - bash script which should (untested) find the
last OMI_SP_yyyymmdd.mat file on the file server, the last MCD43C3 file, and run
read_omno2_v_aug2012.m between those two dates. This will keep the SP mat files
up to date, which can be easily moved on and off the cluster.

MODIFIED FILES:

    Downloading/automodis.py - newline added to end of file, consistent with
Python guidelines.

    Downloading/get_modis.sh -
        1) corrected path to this file's directory to actually point to the
directory, and not this file (only tested on Ubuntu, not Mac OS).
        2) MODDIR now gotten from an environmental variable, this way this
script will not need to be changed if used on different computers.
        3) Date math wrapped in if block to test if OS is Ubuntu or Mac OSX,
which requires different flags for the date command.
        4) Start and end dates passed to automodis now contained in ""; this
keeps them together if there are spaces in the date.
        5) wget commands now execute, instead of being echoed (for debugging)