Automatic run updates and safe cleanup in read_omno2_v_aug2012.m
MODIFIED FILES:
    Downloading/get_modis.sh - clean exit 0 added to end.

    Downloading/get_omno2.sh - clean exit 0 added to end.

    Downloading/run_read_omno2.sh - added paths to data on the file server as
global variables here, with onCluster set to true to trick read_omno2 into using
those
global variables. This helps make it easier to manage paths on multiple
computers.  Also make this run as an interactive shell (#!/bin/bash -i) which
seems to have resolved the problem with the startmatlab alias not working.

    Read_Data/read_omno2_v_aug2012.m - some prints to terminal added for
debugging; the Classes and Utils repos are added to Matlab path. Most
importantly, a cleanup object was added if onCluster is true that will safely
exit if an error occurs.