Partway through making the changes necessary to run a BEHR TEMPO simulation.
I think I've gotten most of the changes set; however, I'd forgotten that since the TEMPO
pixels are smaller than 12 km, I'm going to need to set up the process of matching
profiles to pixels to use interpolation rather than just finding which profile(s) fall in a
given pixel. Since I'd already made this change on the read_wrf_coarse branch, I'm
just going to bring over those specific hunks. This commit is to clean up the working
directory before doing so.

MODIFIED FILES:
     rProfile_WRF.m - modified to read in the profiles for the requested UTC hour, rather than
guessing the hour based on the longitude. Also had to change the path and filename spec
to point to TEMPO files.

    BEHR_main.m - paths pointed to TEMPO and the gridding section removed; gridded for
TEMPO makes no sense because in theory the pixels should overlap perfectly day-to-day
(well at least very close, it may not be perfect).

     read_tempo_simulator.m - corrected the field "RadianceCloudFraction" to the "CloudRadianceFraction"
BEHR was expected. This correction was made after the fact in the TEMPO data file.