Tweaks to make easier to run various a priori in batch and fix parallelization issue
Cherry-picked from the TEMPO branch, there were several conflicts to fix.

MODIFIED FILES:
    BEHR_Main/BEHR_Main.m - now takes "prof_mode" as an additional input which will set
what type of a priori to use. Like the start and end dates, if this is not
passed, the hard coded value will be used.  Also modified the parfor loop to
allow it to run in serial (no workers) if onCluster not defined.

    AMF_tools/rProfile_WRF.m - accepts path to profiles as an input now and
checks that it is a valid directory.

Conflicts:
	AMF_tools/rProfile_WRF.m - did have to remove ref to hour_in;
        otherwise the TEMPO branch mostly won out.
	BEHR_Main/BEHR_main.m - similarly removed utchr references and
        kept normal BEHR filename.
	Utils/TEMPO/read_tempo_simulator.m - not brought over.