Working on adapting the full retrieval to handle any resolution WRF profile.
MODIFIED FILES:
    AMF_tools/rProfiles_WRF.m - checks if the WRF profile it's reading has the
resolution in either direction > 13 km and if so switches to an interpolation
mode, as at those resolutions, there will be pixels that do not have an apriori
profile within their bounds.

    BEHR_Main/BEHR_main.m - sends lat/lon corners to the rProfile_WRF.m instead
of pixel centers; these are necessary for the averaging mode. Also receives the
binning mode back from rProfile_WRF.m (averaging or interpolating) and stores it
in the Data structure as a check that the function is behaving properly.

    Read_Data/read_omno2_v_aug2012.m - added ability to turn off the restart
functionality and fixed a bug with opening a second parallel pool even if it had
already been opened in the runscript.