Four previously uncommitted functions that may be useful across multiple branches.
Cherry pick this commit to bring them in.

NEW FILES:
    Utils/compute_partial_columns.m - estimates partial NO2 columns given a
vector of mixing ratios and pressures.

    Utils/uncross_pix_corners.m - ensures that the given pixel corners are
ordered either clockwise or counter-clockwise and do not cross. This affects
calculation of pixel area.

    WRF_Utils/unstagger.m - unstaggers WRF variables.

    WRF_Utils/wrf_total_cldfra.m - calculates an estimated total cloud fraction
for the given WRF netCDF file, so long as the CLDFRA variable is supplied. Will
append it to that file or write a new file with it appended.