More WIP for unit testing
MODIFIED FILES:
    Read_Data/read_modis_albedo.m - vectorized the lon/lat corners explicitly
instead of relying on the check if any element == 0

    Read_Data/read_omi_sp.m - removed the 0.25 degree edge buffer for the domain;
it was only in the initial cut in v2.1C, not in the final cut.

    Read_Data/read_omno2_v_aug2012.m - transposed Data to a row vector structure
(matches v2.1C shape), also had to fix the temporary corner calculation to use the
whole swath lat/lon and cut down afterwards, or the last scan's corners failed.

    Regrid_tools/Compute_Corner_Pts_ms/compute_edge_pts.m - reverted to v2.1C
with the wrong edge pixel corners: remember, this function will not be used in
the released v3.0A as these computed corners will only be used for unit testing
and the FoV75 and Tiled corners from OMPIXCOR will be provided.

    Utils/h5readomi.m - Apparently there was an extra "end" in this file. Oops.