Upgraded the initial setup process.
NEW FILES:
    setup.sh - bash script that clones the necessary (independent) Git
repos. This seems simpler than dealing with submodules.

MODIFIED FILES:
    BEHR_initial_setup.m - added ability to specify that a path points to
a file, rather than a directory. Which was necessary for the gome2_ler
file, added in this version.

    behr_paths_template.m - added space for the "isfile" structure to be
put, will be a private. Changes ValidatePaths() to use that structure to
decide if it should test whether the path points to a file or a directory.