This squashed together the commits needed to add the ocean mask and BRDF quality flagging
This is a combination of 7 commits.
The first commit's message is:
hdfreadmodis handles SDSs that do not have scale factor or offset attributes

This is the 2nd commit message:

Some changes that support using the MODIS land type-derived ocean mask.

NEW FILES:
    Utils/get_modis_ocean_mask.m - loads the MODIS land type file and converts
it to a binary land/water map.

    Utils/modis_cmg_latlon.m - makes the latitude/longitude grid for MODIS
products on a CMG grid of arbitrary resolution.

MODIFIED FILES:
    BEHR_initial_setup.m - added path to the MODIS land type map.

This is the 3rd commit message:

Fixed hdfreadmodis so that offset = 0 if no offset attribute found

This is the 4th commit message:

modis_cmg_latlon() input syntax made more flexible

MODIFIED FILES:
    Utils/get_modis_ocean_mask.m - updated call to modis_cmg_latlon() to use the
new input syntax.

    Utils/modis_cmg_latlon.m - lonlim and latlim inputs made optional so that
this can be call so as to generate a grid or vectors with the default global lat
and lon limits. Also added help text and better input checking.

This is the 5th commit message:

behr_quality_flags() updated to include clouds in the summary bit and BRDF quality warning

This is the 6th commit message:

For simplicity's sake, modified load_globe_alts to use new behr_paths class

This is the 7th commit message:

Removed reference to the non-existant (on this branch) behr_min_amf_val from behr_quality_flags