Fixed longitude correction in rNmcTmp2
Longitude in the NASA temperature profiles is defined as degrees east of 0,
which requires the OMI data's longitude to be adjusted to match that, as
the OMI data gives west longitudes as negative rather than a positive number
>180.

The old correction of OMI lon + 180 would have worked if the temperature profile
longitude was degrees east of the international date line, but that's not the
case. I have checked this by plotting the surface temperatures for June and
January using Tests/plot_temp_profs.m; with the new lon = mod(lon,360),
in June, you can see the hottest surface temperatures in the Middle East and
India, in Jan, there's a clear match between hot temperatures and Australia.

I've also taken the opportunity to clean up the messiness of rNmcTmp2, specifically,
the interpolation near the end. Using the temperature_prof_unit_test.m, I produced
temperature_prof_test_data.mat (using the old longitude), changed the interpolation
in rNmcTmp2, and used temperature_prof_unit_test again to verify that the same
temperature profiles are obtained.