Supporting changes to name the output files with region and profile mode
MODIFIED FILES:

    BEHR_initial_setup.m - comments for sp_mat_dir and behr_mat_dir updated to
reflect that those directories should be the top of a region/profile mode tree.
Switched mcd43c1 -> mcd43d directory.

    Utils/Constants/behr_filename.m - updated to insert a region and profile
mode into the file name, or wildcards if those aren't given.

    Utils/Constants/behr_paths_template.m - added the SPMatSubdir and
BEHRMatSubdir methods that yield the right path for a given region or region and
profile mode.

    Utils/Constants/sp_savename.m - updated to insert a region into the file
name, or a wildcard if that's not given.