Fix to initial setup for non-standard Python installations and quality flag fix
MODIFIED FILES:
    BEHR_initial_setup.m - instead of building with whatever version of python
is called by the system command "python" this will specifically get the python
executable set in the build-in PYVERSION function, since that ensures that it
will be built by the same version of python as will run it. I made this change
after working with someone who used an Anaconda python installation normally,
but because the GUI Matlab gets a different PATH variable than a program launched
from the Terminal on Macs, it would call the system python by default.

    Utils/behr_quality_flags.m - testing the quality flags with main_priori_tests
in the BEHR-core repo, found that pixels with low-quality MODIS data were not being
set in the quality bit. This has been corrected.