Some changes to allow testing BEHR v3.
MODIFIED FILES:
    AMF_tools/integPr2.m - now returns 0 if the tropopause (i.e. top)
pressure is greater than the surface pressure. This should not affect
the core BEHR code because the calls to it in omiAmfAK2 already do
that, but this way it is consistent.

    Utils/cat_sat_data.m - modified to handle discontinuous date ranges
and to allow you to reject pixels during the concatenation, which is
useful if you want to collect statistics on valid pixels only.