Fix issue where clear scattering weights are all NaNs and cloudy ones are not
MODIFIED FILES:
    AMF_tools/integPr2.m - will give error or warning about NaNs in the partial
columns even if all of the partial columns are NaNs now. The case where all of
the elements are NaNs will be handled in omiAmfAK2 now.

    AMF_tools/omiAmfAK2.m - calculated quantities intialized to NaN instead of
0. Integration skipped if any of the NO2 profiles, clear/cloudy scattering
weights, or temperature correction are all NaNs for a given pixel. This fixes an
issue discovered in data for 2006-01-01 where NaNs in the MODIS albedo caused
the clear sky scattering weights to be all NaNs but the cloudy sky weights were
not, effectively making the clear sky AMF 0. With this change, that case will
produce a NaN for the AMF, correctly indicating that the AMF calculation could
not be carried out.

    Utils/behr_quality_flags.m - when setting the error bit where
XTrackQualityFlags indicates that the pixel is affected by the row anomaly,
changed the test from XTrackFlags ~= 0 to XTrackFlags > 0; the latter test
returns False when XTrackFlags is a NaN, i.e. a fill value. It seems that before
November 2008 (but excluding 2005 possibly) XTrackQualityFlags was set to a fill
value for all pixels. So the XFlags ~= 0 criteria would cause all of 2006-2008
to get rejected, whereas XFlags > 0 does not.